% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{dwc_add_vernacular_gbif}
\alias{dwc_add_vernacular_gbif}
\title{Añadir vernacularName desde GBIF (filtrando por idioma y país)}
\usage{
dwc_add_vernacular_gbif(df, lang = "es", country = "CO", prefer_country = TRUE)
}
\arguments{
\item{df}{data.frame con columna taxonID (GBIF usageKey del aceptado) y scientificName}

\item{lang}{código de idioma (p.ej., "es")}

\item{country}{código de país ISO2 (p.ej., "CO") para preferir nombres de Colombia}

\item{prefer_country}{si TRUE, prioriza nombres con country == CO cuando existan}
}
\description{
Añadir vernacularName desde GBIF (filtrando por idioma y país)
}
