% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{dwc_derive_coords}
\alias{dwc_derive_coords}
\title{Derivar campos verbatim de coordenadas
Crea geodeticDatum = "WGS84" si no existe
Requiere: decimalLatitude, decimalLongitude (WGS84, numéricos o coercibles)}
\usage{
dwc_derive_coords(df)
}
\description{
Derivar campos verbatim de coordenadas
Crea geodeticDatum = "WGS84" si no existe
Requiere: decimalLatitude, decimalLongitude (WGS84, numéricos o coercibles)
}
