% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{dwc_georef_colombia}
\alias{dwc_georef_colombia}
\title{Georreferenciar: asignar stateProvince/municipality/locationID (DANE)}
\usage{
dwc_georef_colombia(df)
}
\arguments{
\item{df}{data.frame con decimalLatitude y decimalLongitude (WGS84)}
}
\description{
Georreferenciar: asignar stateProvince/municipality/locationID (DANE)
}
