% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{dwc_normalize_input}
\alias{dwc_normalize_input}
\title{Normaliza entradas mínimas y deriva year/month/day
Crea columnas opcionales si faltan (language, accessRights, basisOfRecord, geodeticDatum)
Acepta snake_case -> camelCase para mínimos}
\usage{
dwc_normalize_input(df_min)
}
\description{
Normaliza entradas mínimas y deriva year/month/day
Crea columnas opcionales si faltan (language, accessRights, basisOfRecord, geodeticDatum)
Acepta snake_case -> camelCase para mínimos
}
