% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{dwc_require_minimum}
\alias{dwc_require_minimum}
\title{Verifica columnas mínimas requeridas (existencia y no-NA / no-vacías)}
\usage{
dwc_require_minimum(df)
}
\description{
Verifica columnas mínimas requeridas (existencia y no-NA / no-vacías)
}
