% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{dwc_set_defaults}
\alias{dwc_set_defaults}
\title{Definir valores por defecto del proyecto (perfil)}
\usage{
dwc_set_defaults(defaults = list())
}
\arguments{
\item{defaults}{lista con campos DwC (ej. institutionCode, collectionCode, ...)}
}
\value{
la lista de defaults activa (invisible)
}
\description{
Definir valores por defecto del proyecto (perfil)
}
